+++ packages/private-apis/src/implementation.ts
@@ -37,7 +37,6 @@ const CORE_MODULES_USING_PRIVATE_APIS = [
 	'@wordpress/preferences',
 	'@wordpress/reusable-blocks',
 	'@wordpress/rich-text',
-	'@wordpress/rich-text-control',
 	'@wordpress/route',
 	'@wordpress/router',
 	'@wordpress/routes',